+++ DesignToolbox/DesignToolbox/Pages/Components/ContentDisplay/BulletList/BulletListConfigurationView.swift
@@ -33,7 +33,7 @@ final class BulletListConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var unorderedBulletIcon: BulletListUnorderedIcon {
+    @Published var unorderedAsset: BulletListUnorderedAsset {
         didSet { updateCode() }
     }
 
@@ -55,7 +55,7 @@ final class BulletListConfigurationModel: ComponentConfiguration {
         label = String(localized: "app_components_common_label_label")
         levelCount = .one
         bulletType = .unordered
-        unorderedBulletIcon = .bullet
+        unorderedAsset = .bullet
         unorderedBulletIsBranded = true
         textStyle = .bodyLarge
         isBold = true
@@ -68,13 +68,13 @@ final class BulletListConfigurationModel: ComponentConfiguration {
     // MARK: Component Configuration
 
     private var iconPattern: String {
-        switch unorderedBulletIcon {
+        switch unorderedAsset {
         case .bullet:
             ""
         case .tick:
-            "icon: .tick"
-        case .free:
-            "icon: .free(\(Image.defaultImageSample()))"
+            "asset: .tick"
+        case .icon:
+            "asset: .icon(\(Image.defaultImageSample()))"
         }
     }
 
@@ -161,30 +161,30 @@ struct BulletListConfigurationView: View {
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.fixedMedium) {
             VStack(alignment: .leading, spacing: theme.spaces.fixedNone) {
-                OUDSChipPicker(title: "app_components_common_type_label",
+                OUDSChipPicker(title: "app_components_common_type_tech",
                                selection: $configurationModel.bulletType,
                                chips: BulletListType.chips)
 
                 if configurationModel.bulletType == .unordered {
-                    OUDSChipPicker(title: "app_components_bulletList_unorderedIcon_label",
-                                   selection: $configurationModel.unorderedBulletIcon,
-                                   chips: BulletListUnorderedIcon.chips)
+                    OUDSChipPicker(title: "app_components_bulletList_unorderedAsset_tech",
+                                   selection: $configurationModel.unorderedAsset,
+                                   chips: BulletListUnorderedAsset.chips)
 
-                    OUDSSwitchItem("app_components_bulletList_unorderedIconBrandColor_label", isOn: $configurationModel.unorderedBulletIsBranded)
+                    OUDSSwitchItem("app_components_bulletList_unorderedAssetBrandColor_tech", isOn: $configurationModel.unorderedBulletIsBranded)
                 }
 
-                OUDSChipPicker(title: "app_components_bulletList_textStyle_label",
+                OUDSChipPicker(title: "app_components_bulletList_textStyle_tech",
                                selection: $configurationModel.textStyle,
                                chips: OUDSBulletList.TextStyle.chips)
 
-                OUDSSwitchItem("app_components_bulletList_bold_label", isOn: $configurationModel.isBold)
+                OUDSSwitchItem("app_components_bulletList_bold_tech", isOn: $configurationModel.isBold)
 
-                OUDSChipPicker(title: "app_components_bulletList_levelCount_label",
+                OUDSChipPicker(title: "app_components_bulletList_levelCount_tech",
                                selection: $configurationModel.levelCount,
                                chips: BulletListLevelCount.chips)
 
                 DesignToolboxEditContentDisclosure {
-                    DesignToolboxTextField(text: $configurationModel.label, label: "app_components_common_label_label")
+                    DesignToolboxTextField(text: $configurationModel.label, label: "app_components_common_label_tech")
                 }
             }
         }
@@ -201,9 +201,9 @@ enum BulletListLevelCount: CaseIterable, CustomStringConvertible {
         case .one:
             "app_components_bulletList_oneLevel_label"
         case .two:
-            "app_components_bulletList_twoLevel_label"
+            "app_components_bulletList_twoLevels_label"
         case .three:
-            "app_components_bulletList_threeLevel_label"
+            "app_components_bulletList_threeLevels_label"
         }
     }
 
@@ -272,10 +272,10 @@ extension OUDSBulletList.TextStyle: @retroactive CaseIterable, @retroactive Cust
 
 // MARK: - Bullet List Unordered Icon
 
-enum BulletListUnorderedIcon: CaseIterable, CustomStringConvertible {
+enum BulletListUnorderedAsset: CaseIterable, CustomStringConvertible {
     case bullet
     case tick
-    case free
+    case icon
 
     // NOTE: Not localized because it is a technical name
     var description: String {
@@ -284,8 +284,8 @@ enum BulletListUnorderedIcon: CaseIterable, CustomStringConvertible {
             "Bullet"
         case .tick:
             "Tick"
-        case .free:
-            "Free"
+        case .icon:
+            "Icon"
         }
     }