+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/TextInput/TextInputElement.swift
@@ -20,7 +20,7 @@ struct TextInputElement: DesignToolboxElement {
     let pageDescription: AnyView
 
     init() {
-        name = "app_components_textInput_label".localized()
+        name = "app_components_textInput_tech".localized()
         illustration = AnyView(TextInputIllustration())
         pageDescription = AnyView(DesignToolboxElementPage(
             name: name,
@@ -35,7 +35,7 @@ private struct TextInputIllustration: View {
     @State private var text: String = ""
 
     var body: some View {
-        OUDSTextInput(label: "Label",
+        OUDSTextInput(label: "app_components_common_label_label".localized(),
                       text: $text,
                       helperText: String(localized: "app_components_textInputHelperText_label"))
     }