+++ DesignToolbox/DesignToolbox/Pages/Components/Indicators/Tags/TagElements.swift
@@ -25,7 +25,7 @@ struct TagElements: DesignToolboxElement {
             InputTagElement(),
         ]
 
-        name = "app_components_tag_label".localized()
+        name = "app_components_tag_tech".localized()
         illustration = AnyView(TagIllustration())
         pageDescription = AnyView(DesignToolboxElementPage(
             name: name,
@@ -38,6 +38,10 @@ struct TagElements: DesignToolboxElement {
 private struct TagIllustration: View {
 
     var body: some View {
-        OUDSTag(label: "Label", status: .positive(leading: .none), appearance: .emphasized, shape: .rounded, size: .default)
+        OUDSTag(label: "app_components_common_label_label".localized(),
+                status: .positive(leading: .none),
+                appearance: .emphasized,
+                shape: .rounded,
+                size: .default)
     }
 }